Update dependencies to latest tags and switch to hex packages where possible
All dependecies are now pinned to a specific version or commit and
the dependencies will be fetched from hex.pm where possible. This change
implicates that erlang.mk refactors the rebar.config to a rebar3
specific one, but backward compatibility for rebar2 projects will be
ensured through rebar.config.script.