Add `jsx` to `.app.src` to include it in release builds.
While using this library (via `slackerl`) I found crash logs in production
with `{undef,[{jsx,decode,...` indicating the `jsx` isn't present. Sure enough
it wasn't included in the release `lib/` directory because nothing was requiring it.

In this patch I'm adding the dependency into the `.app.src` list of applications so
that it will be properly included in releases and initialize if it needs to. This was
hidden at dev time because the library was available locally.

I've added a path-level version bump since this is fixing what I believe is a simple
bug and leaving all behaviors and interfaces otherwise untouched.